move block methods to file
